Move aerospike mocks to internal package
