+++ receiver/aerospikereceiver/cluster/cluster_test.go
@@ -9,7 +9,7 @@ import (
 	as "github.com/aerospike/aerospike-client-go/v8"
 	"github.com/stretchr/testify/require"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/aerospikereceiver/cluster/mocks"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/aerospikereceiver/internal/cluster/mocks"
 )
 
 func TestCluster_GetNodes(t *testing.T) {
