+++ receiver/aerospikereceiver/cluster/node_test.go
@@ -10,7 +10,7 @@ import (
 	as "github.com/aerospike/aerospike-client-go/v8"
 	"github.com/stretchr/testify/require"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/aerospikereceiver/cluster/mocks"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/aerospikereceiver/internal/cluster/mocks"
 )
 
 func TestNode_NewConnNode(t *testing.T) {
