+++ receiver/aerospikereceiver/internal/cluster/mocks/Node.go
