+++ receiver/aerospikereceiver/internal/cluster/mocks/asclient.go
